Added mock actions to aid us as we implement cards
Resource checker completely works, it builds a graph, the pruning requirements are within the function itself which seems clunky. It also handles all of the xor decisions, which seems like the best way to handle it. Right now the game is totally implemented other than the cards and scoring ( I mean i guess im probably missing a few holes ).

Next up is to implement the resources (play card gain resource).